meta: split the meta module
